refactor: split off craftability from calculator
The calculator should not manage
listeners on its own.
