+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/calculation/CalculationException.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.api.autocrafting.calculation;
 
-class CalculationException extends RuntimeException {
+public class CalculationException extends RuntimeException {
     protected CalculationException(final String message) {
         super(message);
     }
