+++ refinedstorage-autocrafting-api/src/test/java/com/refinedmods/refinedstorage/api/autocrafting/craftability/CraftabilityTest.java
@@ -1,6 +1,9 @@
-package com.refinedmods.refinedstorage.api.autocrafting.calculation;
+package com.refinedmods.refinedstorage.api.autocrafting.craftability;
 
 import com.refinedmods.refinedstorage.api.autocrafting.PatternRepository;
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CancellationToken;
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CraftingCalculator;
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CraftingCalculatorImpl;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
 
@@ -14,9 +17,10 @@ import static com.refinedmods.refinedstorage.api.autocrafting.PatternBuilder.pat
 import static com.refinedmods.refinedstorage.api.autocrafting.ResourceFixtures.CRAFTING_TABLE;
 import static com.refinedmods.refinedstorage.api.autocrafting.ResourceFixtures.OAK_LOG;
 import static com.refinedmods.refinedstorage.api.autocrafting.ResourceFixtures.OAK_PLANKS;
+import static com.refinedmods.refinedstorage.api.autocrafting.craftability.IsCraftableCraftingCalculatorListener.binarySearchMaxAmount;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class CraftingCalculatorImplTest {
+class CraftabilityTest {
     @Test
     void shouldNotFindMaxAmountIfThereAreAlwaysMissingResources() {
         // Arrange
@@ -36,7 +40,7 @@ class CraftingCalculatorImplTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final long maxAmount = sut.getMaxAmount(CRAFTING_TABLE, CancellationToken.NONE);
+        final long maxAmount = binarySearchMaxAmount(sut, CRAFTING_TABLE, CancellationToken.NONE);
 
         // Assert
         assertThat(maxAmount).isZero();
@@ -62,7 +66,7 @@ class CraftingCalculatorImplTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final long maxAmount = sut.getMaxAmount(CRAFTING_TABLE, CancellationToken.NONE);
+        final long maxAmount = binarySearchMaxAmount(sut, CRAFTING_TABLE, CancellationToken.NONE);
 
         // Assert
         assertThat(maxAmount).isEqualTo(amountPossible);
@@ -87,7 +91,7 @@ class CraftingCalculatorImplTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final long maxAmount = sut.getMaxAmount(CRAFTING_TABLE, CancellationToken.NONE);
+        final long maxAmount = binarySearchMaxAmount(sut, CRAFTING_TABLE, CancellationToken.NONE);
 
         // Assert
         assertThat(maxAmount).isZero();
